Switch to CSV encoding
The MQTT library has a default message limit of 128 bytes. In order to
keep messages within this limit it was necessary to switch to a more
compact representation: CSV. In case this is still too large we could
even switch to a binary representation.

In addition the client name is replaced with a client ID to save some
more space.

Further changes:
- Remove co2SensorStartupTime and use uptime instead
- Report temperature in Kelvin so that we can check represent missing
  values consistently as 0
- Report uptime to detect reboots